Split out BasicWaylandSurfaceObserver
