+++ src/server/frontend_wayland/CMakeLists.txt
@@ -7,6 +7,7 @@ include_directories(${mirwayland_includes})
 set(
   WAYLAND_SOURCES
 
+  basic_wayland_surface_observer.cpp basic_wayland_surface_observer.h
   mir_display.cpp               mir_display.h
   wayland_default_configuration.cpp
   wayland_connector.cpp         wayland_connector.h
